+++ lib/Payever/ExternalIntegration/Core/Exception/ConfigurationException.php
@@ -16,5 +16,4 @@ namespace Payever\ExternalIntegration\Core\Exception;
 
 class ConfigurationException extends \InvalidArgumentException
 {
-
 }