+++ lib/Payever/ExternalIntegration/Core/Exception/PayeverCommunicationException.php
@@ -16,5 +16,4 @@ namespace Payever\ExternalIntegration\Core\Exception;
 
 class PayeverCommunicationException extends \UnexpectedValueException
 {
-
 }